[Feat] move Activity to main folder
